Move URL building to RequestEncoderContainer
